Pull the latest dataset before trying to update it
Though this is a kinda expensive operation (the entire dataset needs to be re-parsed and loaded into memory, maybe I should look into just storing video IDs), it better keeps the server safe as a precaution.

Ideally, with this server being the only party updating the dataset, this wouldn't be necessary. However, it's good to be safe.