Re-add github to known hosts?
Something strange is going on. In the `setup-git.sh` script, we make sure to set up SSH by adding the private key to the key agent and then adding GitHub to the known hosts lists. There, testing SSH works.

Once we get to the commit script, it seems like this all wears off and host key verification fails all over again. I'm not sure why this is the case, but this will check to see if we need to re-add GitHub back to the list of known hosts (I'd rather know why this is the case than using this as a stopgap however).