Add refresh-dataset script
The auto-update failed today because I had committed changes without triggering a deploy. The deployment was then out of sync, causing the update to fail.

I need:
1. A way to easily and quickly re-sync the dataset when reviews are missed.
2. To pull latest changes before committing a new review.

The second point there is complicated unless we assume that only these scripts should ever modify the dataset. The review processing needs an up-to-date dataset of all TND videos to avoid duplicates in the set, and if changes are somehow made to that outside of the callback server's work or syncing the datasets, then there may be issues since the server loads that data in memory upon launch. It would need to update the changes and reload the videos in memory which is a little expensive.

If we can trust that the datasets are only modified by syncing and the server pulling new changes, keeping itself in sync, then we can safely pull whenever we receive a video or in the commit script AFTER adding the review.