Pull changes before updating dataset.
There seems to be an issue where even if no other external changes are made to the repo which could cause the callback server to go out of date, it can still end up out of date after one successful push.

I'm not sure if this will fix any underlying issue, but it will at least account for changes made to other non-dataset files in the repo.