Make server synchronous
This is not a good long term solution since the recommendation is to only use express middleware in fastify as a temporary solution, but this was the easiest way to make responses to requests synchronous using a queue. The issue here was that multiple requests were being processed at the same time, causing strange behavior with the local dataset state on the deployment.

This change ensures that only a single request can run and modify the dataset at a single time, helping the callback server stay consistent.