Add `collect` reducer
This will be used to collect parsed reviews into a list of review rows which can be output to a CSV.

Currently the strategy for doing this is quite messy, where the parsed reviews types are just members of a discriminated union. Based upon the `type` field, the right collection strategy is delegated to.

A possible cleaner solution in TS would be to parse reviews into classes, then rely on dynamic dispatch or possibly a visitor pattern to collect the reviews.